package into duke
