+++ build.gradle
@@ -29,7 +29,7 @@ test {
 }
 
 application {
-    mainClass.set("MentalHealth")
+    mainClass.set("duke.MentalHealth")
 }
 
 shadowJar {
