+++ data/duke.txt
@@ -2,3 +2,4 @@ T | 0 | borrow book
 D | 0 | return book | 2/12/2019 1800
 T | 0 | borrow book
 T | 0 | study
+D | 0 | return book | 2/12/2019 1800
