+++ src/main/java/duke/Event.java
@@ -1,6 +1,4 @@
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
+package duke;
 
 
 public class Event extends IndividualTask{
