+++ src/main/java/duke/MentalHealthException.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class MentalHealthException extends Exception {
     public MentalHealthException(String message) {
         super(message);
