+++ src/test/java/duke/ToDoTest.java
@@ -1,4 +1,6 @@
+package duke;
 
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -8,13 +10,13 @@ public class ToDoTest {
     public void testSaveToFileFormat() {
         ToDo todo = new ToDo("Finish homework");
         String expected = "T | 0 | Finish homework";
-        assertEquals(expected, todo.saveToFileFormat());
+        Assertions.assertEquals(expected, todo.saveToFileFormat());
     }
 
     @Test
     public void testToString() {
         ToDo todo = new ToDo("Finish homework");
         String expected = "[T][ ] Finish homework";
-        assertEquals(expected, todo.toString());
+        Assertions.assertEquals(expected, todo.toString());
     }
 }
